requirements.txt
numpy
opencv-python
open3d
pillow
pyyaml
exif
scipy
tqdm